+++ system/src/Grav/Common/Form/FormFlash.php
@@ -10,54 +10,10 @@
 namespace Grav\Common\Form;
 
 use Grav\Common\Filesystem\Folder;
-use Grav\Common\Grav;
-use Grav\Common\User\Interfaces\UserInterface;
-use RocketTheme\Toolbox\File\YamlFile;
-use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
+use Grav\Framework\Form\FormFlash as FrameworkFormFlash;
 
-class FormFlash extends \Grav\Framework\Form\FormFlash
+class FormFlash extends FrameworkFormFlash
 {
-    /**
-     * @param string $sessionId
-     */
-    public static function clearSession(string $sessionId): void
-    {
-        $folder = static::getSessionTmpDir($sessionId);
-        if (is_dir($folder)) {
-            Folder::delete($folder);
-        }
-    }
-
-    /**
-     * @param string $sessionId
-     * @return string
-     */
-    public static function getSessionTmpDir(string $sessionId): string
-    {
-        /** @var UniformResourceLocator $locator */
-        $locator = Grav::instance()['locator'];
-
-        return $locator->findResource("tmp://forms/{$sessionId}", true, true);
-    }
-
-    /**
-     * @param UserInterface|null $user
-     * @return $this
-     */
-    public function setUser(UserInterface $user = null)
-    {
-        if ($user && $user->username) {
-            $this->user = [
-                'username' => $user->username,
-                'email' => $user->email ?? ''
-            ];
-        } else {
-            $this->user = null;
-        }
-
-        return $this;
-    }
-
     /**
      * @return array
      * @deprecated 1.6 For backwards compatibility only, do not use
@@ -89,8 +45,11 @@ class FormFlash extends \Grav\Framework\Form\FormFlash
      */
     public function uploadFile(string $field, string $filename, array $upload): bool
     {
-        $tmp_dir = $this->getTmpDir();
+        if (!$this->uniqueId) {
+            return false;
+        }
 
+        $tmp_dir = $this->getTmpDir();
         Folder::create($tmp_dir);
 
         $tmp_file = $upload['file']['tmp_name'];
@@ -118,8 +77,11 @@ class FormFlash extends \Grav\Framework\Form\FormFlash
      */
     public function cropFile(string $field, string $filename, array $upload, array $crop): bool
     {
-        $tmp_dir = $this->getTmpDir();
+        if (!$this->uniqueId) {
+            return false;
+        }
 
+        $tmp_dir = $this->getTmpDir();
         Folder::create($tmp_dir);
 
         $tmp_file = $upload['file']['tmp_name'];
@@ -135,33 +97,5 @@ class FormFlash extends \Grav\Framework\Form\FormFlash
         $this->addFileInternal($field, $filename, $upload['file'], $crop);
 
         return true;
-    }
-
-    /**
-     * @return YamlFile
-     */
-    protected function getTmpIndex(): YamlFile
-    {
-        // Do not use CompiledYamlFile as the file can change multiple times per second.
-        return YamlFile::instance($this->getTmpDir() . '/index.yaml');
-    }
-
-    /**
-     * @param string $name
-     */
-    protected function removeTmpFile(string $name): void
-    {
-        $filename = $this->getTmpDir() . '/' . $name;
-        if ($name && is_file($filename)) {
-            unlink($filename);
-        }
-    }
-
-    protected function removeTmpDir(): void
-    {
-        $tmpDir = $this->getTmpDir();
-        if (file_exists($tmpDir)) {
-            Folder::delete($tmpDir);
-        }
     }
 }