+++ system/src/Grav/Framework/Form/FormFlash.php
@@ -10,12 +10,18 @@
 namespace Grav\Framework\Form;
 
 use Grav\Common\Filesystem\Folder;
+use Grav\Common\Grav;
+use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Utils;
+use Grav\Framework\Form\Interfaces\FormFlashInterface;
 use Psr\Http\Message\UploadedFileInterface;
 use RocketTheme\Toolbox\File\YamlFile;
+use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
-class FormFlash implements \JsonSerializable
+class FormFlash implements FormFlashInterface
 {
+    /** @var bool */
+    protected $exists;
     /** @var string */
     protected $sessionId;
     /** @var string */
@@ -26,6 +32,10 @@ class FormFlash implements \JsonSerializable
     protected $url;
     /** @var array */
     protected $user;
+    /** @var int */
+    protected $createdTimestamp;
+    /** @var int */
+    protected $updatedTimestamp;
     /** @var array */
     protected $data;
     /** @var array */
@@ -34,28 +44,35 @@ class FormFlash implements \JsonSerializable
     protected $uploadedFiles;
     /** @var string[] */
     protected $uploadObjects;
-    /** @var bool */
-    protected $exists;
+    /** @var string|null */
+    protected $folder;
 
     /**
-     * @param string $sessionId
-     * @return string
+     * @inheritDoc
      */
-    public static function getSessionTmpDir(string $sessionId): string
+    public function __construct($config)
     {
-        return "tmp://forms/{$sessionId}";
-    }
+        // Backwards compatibility with Grav 1.6 plugins.
+        if (!is_array($config)) {
+            user_error(__CLASS__ . '::' . __FUNCTION__ . '($sessionId, $uniqueId, $formName) is deprecated since Grav 1.6.11, use $config parameter instead', E_USER_DEPRECATED);
 
-    /**
-     * FormFlashObject constructor.
-     * @param string $sessionId
-     * @param string $uniqueId
-     * @param string|null $formName
-     */
-    public function __construct(string $sessionId, string $uniqueId, string $formName = null)
-    {
-        $this->sessionId = $sessionId;
-        $this->uniqueId = $uniqueId;
+            $args = func_get_args();
+            $config = [
+                'session_id' => $args[0],
+                'unique_id' => $args[1] ?? null,
+                'form_name' => $args[2] ?? null,
+            ];
+        }
+
+        $this->sessionId = $config['session_id'] ?? 'no-session';
+        $this->uniqueId = $config['unique_id'] ?? '';
+        $this->folder = $config['folder'] ?? 'tmp://forms';
+
+        /** @var UniformResourceLocator $locator */
+        $locator = Grav::instance()['locator'];
+        if ($locator->isStream($this->folder)) {
+            $this->folder = $locator->findResource($this->folder, true, true);
+        }
 
         $file = $this->getTmpIndex();
         $this->exists = $file->exists();
@@ -66,65 +83,58 @@ class FormFlash implements \JsonSerializable
             } catch (\Exception $e) {
                 $data = [];
             }
-            $this->formName = null !== $formName ? $content['form'] ?? '' : '';
+            $this->formName = $content['form'] ?? $config['form_name'] ?? '';
             $this->url = $data['url'] ?? '';
             $this->user = $data['user'] ?? null;
+            $this->updatedTimestamp = $data['timestamps']['updated'] ?? time();
+            $this->createdTimestamp = $data['timestamps']['created'] ?? $this->updatedTimestamp;
             $this->data = $data['data'] ?? null;
             $this->files = $data['files'] ?? [];
         } else {
-            $this->formName = $formName;
+            $this->formName = $config['form_name'] ?? '';
             $this->url = '';
+            $this->createdTimestamp = $this->updatedTimestamp = time();
             $this->files = [];
         }
     }
 
     /**
-     * @return string
+     * @inheritDoc
+     */
+    public function getSessionId(): string
+    {
+        return $this->sessionId;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getUniqueId(): string
+    {
+        return $this->uniqueId;
+    }
+
+    /**
+     * @deprecated 1.6.11 Use '->getUniqueId()' method instead.
+     */
+    public function getUniqieId(): string
+    {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.6.11, use ->getUniqueId() method instead', E_USER_DEPRECATED);
+
+        return $this->getUniqueId();
+    }
+
+    /**
+     * @inheritDoc
      */
     public function getFormName(): string
     {
         return $this->formName;
     }
 
-    /**
-     * @return string
-     */
-    public function getUniqieId(): string
-    {
-        return $this->uniqueId;
-    }
 
     /**
-     * @return bool
-     */
-    public function exists(): bool
-    {
-        return $this->exists;
-    }
-
-    /**
-     * @return $this
-     */
-    public function save(): self
-    {
-        $file = $this->getTmpIndex();
-        $file->save($this->jsonSerialize());
-        $this->exists = true;
-
-        return $this;
-    }
-
-    public function delete(): self
-    {
-        $this->removeTmpDir();
-        $this->files = [];
-        $this->exists = false;
-
-        return $this;
-    }
-
-    /**
-     * @return string
+     * @inheritDoc
      */
     public function getUrl(): string
     {
@@ -132,18 +142,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * @param string $url
-     * @return $this
-     */
-    public function setUrl(string $url): self
-    {
-        $this->url = $url;
-
-        return $this;
-    }
-
-    /**
-     * @return string
+     * @inheritDoc
      */
     public function getUsername(): string
     {
@@ -151,7 +150,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * @return string
+     * @inheritDoc
      */
     public function getUserEmail(): string
     {
@@ -159,40 +158,84 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * @param string|null $username
-     * @return $this
+     * @inheritDoc
      */
-    public function setUserName(string $username = null): self
+    public function getCreatedTimestamp(): int
     {
-        $this->user['username'] = $username;
-
-        return $this;
+        return $this->createdTimestamp;
     }
 
     /**
-     * @param string|null $email
-     * @return $this
+     * @inheritDoc
      */
-    public function setUserEmail(string $email = null): self
+    public function getUpdatedTimestamp(): int
     {
-        $this->user['email'] = $email;
-
-        return $this;
+        return $this->updatedTimestamp;
     }
 
+
+    /**
+     * @inheritDoc
+     */
     public function getData(): ?array
     {
         return $this->data;
     }
 
+    /**
+     * @inheritDoc
+     */
     public function setData(?array $data): void
     {
         $this->data = $data;
     }
 
     /**
-     * @param string $field
-     * @return array
+     * @inheritDoc
+     */
+    public function exists(): bool
+    {
+        return $this->exists;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function save(): self
+    {
+        if (!$this->sessionId && $this->uniqueId) {
+            return $this;
+        }
+
+        if ($this->data || $this->files) {
+            // Only save if there is data or files to be saved.
+            $file = $this->getTmpIndex();
+            $file->save($this->jsonSerialize());
+            $this->exists = true;
+        } elseif ($this->exists) {
+            // Delete empty form flash if it exists (it carries no information).
+            return $this->delete();
+        }
+
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function delete(): self
+    {
+        if ($this->sessionId && $this->uniqueId) {
+            $this->removeTmpDir();
+            $this->files = [];
+            $this->exists = false;
+        }
+
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
      */
     public function getFilesByField(string $field): array
     {
@@ -208,8 +251,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * @param bool $includeOriginal
-     * @return array
+     * @inheritDoc
      */
     public function getFilesByFields($includeOriginal = false): array
     {
@@ -225,12 +267,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * Add uploaded file to the form flash.
-     *
-     * @param UploadedFileInterface $upload
-     * @param string|null $field
-     * @param array|null $crop
-     * @return string Return name of the file
+     * @inheritDoc
      */
     public function addUploadedFile(UploadedFileInterface $upload, string $field = null, array $crop = null): string
     {
@@ -256,12 +293,7 @@ class FormFlash implements \JsonSerializable
 
 
     /**
-     * Add existing file to the form flash.
-     *
-     * @param string $filename
-     * @param string $field
-     * @param array $crop
-     * @return bool
+     * @inheritDoc
      */
     public function addFile(string $filename, string $field, array $crop = null): bool
     {
@@ -282,11 +314,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * Remove any file from form flash.
-     *
-     * @param string $name
-     * @param string $field
-     * @return bool
+     * @inheritDoc
      */
     public function removeFile(string $name, string $field = null): bool
     {
@@ -318,7 +346,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * Clear form flash from all uploaded files.
+     * @inheritDoc
      */
     public function clearFiles()
     {
@@ -332,7 +360,7 @@ class FormFlash implements \JsonSerializable
     }
 
     /**
-     * @return array
+     * @inheritDoc
      */
     public function jsonSerialize(): array
     {
@@ -341,9 +369,64 @@ class FormFlash implements \JsonSerializable
             'unique_id' => $this->uniqueId,
             'url' => $this->url,
             'user' => $this->user,
+            'timestamps' => [
+                'created' => $this->createdTimestamp,
+                'updated' => time(),
+            ],
             'data' => $this->data,
             'files' => $this->files
         ];
+    }
+
+    /**
+     * @param string $url
+     * @return $this
+     */
+    public function setUrl(string $url): self
+    {
+        $this->url = $url;
+
+        return $this;
+    }
+
+    /**
+     * @param UserInterface|null $user
+     * @return $this
+     */
+    public function setUser(UserInterface $user = null)
+    {
+        if ($user && $user->username) {
+            $this->user = [
+                'username' => $user->username,
+                'email' => $user->email ?? ''
+            ];
+        } else {
+            $this->user = null;
+        }
+
+        return $this;
+    }
+
+    /**
+     * @param string|null $username
+     * @return $this
+     */
+    public function setUserName(string $username = null): self
+    {
+        $this->user['username'] = $username;
+
+        return $this;
+    }
+
+    /**
+     * @param string|null $email
+     * @return $this
+     */
+    public function setUserEmail(string $email = null): self
+    {
+        $this->user['email'] = $email;
+
+        return $this;
     }
 
     /**
@@ -351,7 +434,7 @@ class FormFlash implements \JsonSerializable
      */
     public function getTmpDir(): string
     {
-        return static::getSessionTmpDir($this->sessionId) . '/' . $this->uniqueId;
+        return $this->sessionId && $this->uniqueId ? "{$this->folder}/{$this->sessionId}/{$this->uniqueId}" : '';
     }
 
     /**
@@ -368,8 +451,9 @@ class FormFlash implements \JsonSerializable
      */
     protected function removeTmpFile(string $name): void
     {
-        $filename = $this->getTmpDir() . '/' . $name;
-        if ($name && is_file($filename)) {
+        $tmpDir = $this->getTmpDir();
+        $filename =  $tmpDir ? $tmpDir . '/' . $name : '';
+        if ($name && $filename && is_file($filename)) {
             unlink($filename);
         }
     }
@@ -377,7 +461,7 @@ class FormFlash implements \JsonSerializable
     protected function removeTmpDir(): void
     {
         $tmpDir = $this->getTmpDir();
-        if (file_exists($tmpDir)) {
+        if ($tmpDir && file_exists($tmpDir)) {
             Folder::delete($tmpDir);
         }
     }
@@ -390,6 +474,10 @@ class FormFlash implements \JsonSerializable
      */
     protected function addFileInternal(?string $field, string $name, array $data, array $crop = null): void
     {
+        if (!$this->sessionId || !$this->uniqueId) {
+            throw new \RuntimeException('Cannot upload files: unique id not defined');
+        }
+
         $field = $field ?: 'undefined';
         if (!isset($this->files[$field])) {
             $this->files[$field] = [];